Cleanup & rename fixes
