--- db/models/provider.go
@@ -1,14 +1,0 @@
-package models
-
-type Provider int
-
-const (
-	SUKEBEI Provider = iota + 1
-	NYAA
-	YTS
-	LIME
-	SUBSPLEASE
-	MANGADEX
-	WEBTOON
-	DYNASTY
-)
