--- db/models/subscriptions.go
@@ -1,10 +1,0 @@
-package models
-
-type Subscriptions interface {
-	All() ([]Subscription, error)
-	Get(uint) (*Subscription, error)
-
-	New(Subscription) (*Subscription, error)
-	Update(Subscription) error
-	Delete(uint) error
-}
